+++ code/packages/rust/ruby-parser/src/_grammar.rs
@@ -393,12 +393,18 @@ pub fn parser_grammar() -> ParserGrammar {
                         GrammarElement::Literal { value: r#"**"#.to_string() },
                     ] }) },
                 GrammarElement::TokenReference { name: r#"NAME"#.to_string() },
-                GrammarElement::Optional { element: Box::new(GrammarElement::Sequence { elements: vec![
-                        GrammarElement::TokenReference { name: r#"EQUALS"#.to_string() },
-                        GrammarElement::RuleReference { name: r#"expression"#.to_string() },
+                GrammarElement::Optional { element: Box::new(GrammarElement::Alternation { choices: vec![
+                        GrammarElement::Sequence { elements: vec![
+                            GrammarElement::TokenReference { name: r#"COLON"#.to_string() },
+                            GrammarElement::Optional { element: Box::new(GrammarElement::RuleReference { name: r#"expression"#.to_string() }) },
+                        ] },
+                        GrammarElement::Sequence { elements: vec![
+                            GrammarElement::TokenReference { name: r#"EQUALS"#.to_string() },
+                            GrammarElement::RuleReference { name: r#"expression"#.to_string() },
+                        ] },
                     ] }) },
             ] },
-            line_number: 280,
+            line_number: 312,
         },
         GrammarRule {
             name: r#"if_statement"#.to_string(),
@@ -415,7 +421,7 @@ pub fn parser_grammar() -> ParserGrammar {
                 GrammarElement::Optional { element: Box::new(GrammarElement::RuleReference { name: r#"else_clause"#.to_string() }) },
                 GrammarElement::Literal { value: r#"end"#.to_string() },
             ] },
-            line_number: 281,
+            line_number: 313,
         },
         GrammarRule {
             name: r#"elsif_clause"#.to_string(),
@@ -429,7 +435,7 @@ pub fn parser_grammar() -> ParserGrammar {
                         GrammarElement::RuleReference { name: r#"statement"#.to_string() },
                     ] }) },
             ] },
-            line_number: 282,
+            line_number: 314,
         },
         GrammarRule {
             name: r#"else_clause"#.to_string(),
@@ -440,7 +446,7 @@ pub fn parser_grammar() -> ParserGrammar {
                         GrammarElement::RuleReference { name: r#"statement"#.to_string() },
                     ] }) },
             ] },
-            line_number: 283,
+            line_number: 315,
         },
         GrammarRule {
             name: r#"unless_statement"#.to_string(),
@@ -455,7 +461,7 @@ pub fn parser_grammar() -> ParserGrammar {
                 GrammarElement::Optional { element: Box::new(GrammarElement::RuleReference { name: r#"else_clause"#.to_string() }) },
                 GrammarElement::Literal { value: r#"end"#.to_string() },
             ] },
-            line_number: 284,
+            line_number: 316,
         },
         GrammarRule {
             name: r#"while_statement"#.to_string(),
@@ -468,7 +474,7 @@ pub fn parser_grammar() -> ParserGrammar {
                     ] }) },
                 GrammarElement::Literal { value: r#"end"#.to_string() },
             ] },
-            line_number: 285,
+            line_number: 317,
         },
         GrammarRule {
             name: r#"until_statement"#.to_string(),
@@ -481,7 +487,7 @@ pub fn parser_grammar() -> ParserGrammar {
                     ] }) },
                 GrammarElement::Literal { value: r#"end"#.to_string() },
             ] },
-            line_number: 286,
+            line_number: 318,
         },
         GrammarRule {
             name: r#"case_statement"#.to_string(),
@@ -495,7 +501,7 @@ pub fn parser_grammar() -> ParserGrammar {
                 GrammarElement::Optional { element: Box::new(GrammarElement::RuleReference { name: r#"else_clause"#.to_string() }) },
                 GrammarElement::Literal { value: r#"end"#.to_string() },
             ] },
-            line_number: 309,
+            line_number: 341,
         },
         GrammarRule {
             name: r#"when_clause"#.to_string(),
@@ -514,7 +520,7 @@ pub fn parser_grammar() -> ParserGrammar {
                         GrammarElement::RuleReference { name: r#"statement"#.to_string() },
                     ] }) },
             ] },
-            line_number: 310,
+            line_number: 342,
         },
         GrammarRule {
             name: r#"in_clause"#.to_string(),
@@ -529,7 +535,7 @@ pub fn parser_grammar() -> ParserGrammar {
                         GrammarElement::RuleReference { name: r#"statement"#.to_string() },
                     ] }) },
             ] },
-            line_number: 332,
+            line_number: 364,
         },
         GrammarRule {
             name: r#"pattern"#.to_string(),
@@ -541,7 +547,7 @@ pub fn parser_grammar() -> ParserGrammar {
                 GrammarElement::RuleReference { name: r#"literal_pattern"#.to_string() },
                 GrammarElement::RuleReference { name: r#"binding_pattern"#.to_string() },
             ] },
-            line_number: 333,
+            line_number: 365,
         },
         GrammarRule {
             name: r#"literal_pattern"#.to_string(),
@@ -551,12 +557,12 @@ pub fn parser_grammar() -> ParserGrammar {
                 GrammarElement::RuleReference { name: r#"symbol_literal"#.to_string() },
                 GrammarElement::TokenReference { name: r#"KEYWORD"#.to_string() },
             ] },
-            line_number: 334,
+            line_number: 366,
         },
         GrammarRule {
             name: r#"binding_pattern"#.to_string(),
             body: GrammarElement::TokenReference { name: r#"NAME"#.to_string() },
-            line_number: 335,
+            line_number: 367,
         },
         GrammarRule {
             name: r#"array_pattern"#.to_string(),
@@ -577,7 +583,7 @@ pub fn parser_grammar() -> ParserGrammar {
                     ] }) },
                 GrammarElement::TokenReference { name: r#"RBRACKET"#.to_string() },
             ] },
-            line_number: 336,
+            line_number: 368,
         },
         GrammarRule {
             name: r#"hash_pattern"#.to_string(),
@@ -592,7 +598,7 @@ pub fn parser_grammar() -> ParserGrammar {
                     ] }) },
                 GrammarElement::TokenReference { name: r#"RBRACE"#.to_string() },
             ] },
-            line_number: 337,
+            line_number: 369,
         },
         GrammarRule {
             name: r#"hash_pattern_pair"#.to_string(),
@@ -601,7 +607,7 @@ pub fn parser_grammar() -> ParserGrammar {
                 GrammarElement::TokenReference { name: r#"COLON"#.to_string() },
                 GrammarElement::Optional { element: Box::new(GrammarElement::RuleReference { name: r#"pattern"#.to_string() }) },
             ] },
-            line_number: 338,
+            line_number: 370,
         },
         GrammarRule {
             name: r#"splat_pattern"#.to_string(),
@@ -609,7 +615,7 @@ pub fn parser_grammar() -> ParserGrammar {
                 GrammarElement::Literal { value: r#"*"#.to_string() },
                 GrammarElement::Optional { element: Box::new(GrammarElement::TokenReference { name: r#"NAME"#.to_string() }) },
             ] },
-            line_number: 345,
+            line_number: 377,
         },
         GrammarRule {
             name: r#"pin_pattern"#.to_string(),
@@ -617,7 +623,7 @@ pub fn parser_grammar() -> ParserGrammar {
                 GrammarElement::Literal { value: r#"^"#.to_string() },
                 GrammarElement::TokenReference { name: r#"NAME"#.to_string() },
             ] },
-            line_number: 350,
+            line_number: 382,
         },
         GrammarRule {
             name: r#"class_pattern"#.to_string(),
@@ -633,7 +639,7 @@ pub fn parser_grammar() -> ParserGrammar {
                     ] }) },
                 GrammarElement::TokenReference { name: r#"RPAREN"#.to_string() },
             ] },
-            line_number: 356,
+            line_number: 388,
         },
         GrammarRule {
             name: r#"begin_statement"#.to_string(),
@@ -649,7 +655,7 @@ pub fn parser_grammar() -> ParserGrammar {
                 GrammarElement::Optional { element: Box::new(GrammarElement::RuleReference { name: r#"ensure_clause"#.to_string() }) },
                 GrammarElement::Literal { value: r#"end"#.to_string() },
             ] },
-            line_number: 377,
+            line_number: 409,
         },
         GrammarRule {
             name: r#"rescue_clause"#.to_string(),
@@ -667,7 +673,7 @@ pub fn parser_grammar() -> ParserGrammar {
                         GrammarElement::RuleReference { name: r#"statement"#.to_string() },
                     ] }) },
             ] },
-            line_number: 386,
+            line_number: 418,
         },
         GrammarRule {
             name: r#"exception_list"#.to_string(),
@@ -678,7 +684,7 @@ pub fn parser_grammar() -> ParserGrammar {
                         GrammarElement::TokenReference { name: r#"NAME"#.to_string() },
                     ] }) },
             ] },
-            line_number: 387,
+            line_number: 419,
         },
         GrammarRule {
             name: r#"ensure_clause"#.to_string(),
@@ -689,7 +695,7 @@ pub fn parser_grammar() -> ParserGrammar {
                         GrammarElement::RuleReference { name: r#"statement"#.to_string() },
                     ] }) },
             ] },
-            line_number: 388,
+            line_number: 420,
         },
         GrammarRule {
             name: r#"assignment"#.to_string(),
@@ -713,7 +719,7 @@ pub fn parser_grammar() -> ParserGrammar {
                     ] }) },
                 GrammarElement::RuleReference { name: r#"expression"#.to_string() },
             ] },
-            line_number: 408,
+            line_number: 440,
         },
         GrammarRule {
             name: r#"rightward_assignment"#.to_string(),
@@ -722,7 +728,7 @@ pub fn parser_grammar() -> ParserGrammar {
                 GrammarElement::Literal { value: r#"=>"#.to_string() },
                 GrammarElement::TokenReference { name: r#"NAME"#.to_string() },
             ] },
-            line_number: 427,
+            line_number: 459,
         },
         GrammarRule {
             name: r#"method_call"#.to_string(),
@@ -742,7 +748,7 @@ pub fn parser_grammar() -> ParserGrammar {
                 GrammarElement::TokenReference { name: r#"RPAREN"#.to_string() },
                 GrammarElement::Repetition { element: Box::new(GrammarElement::RuleReference { name: r#"dot_call"#.to_string() }) },
             ] },
-            line_number: 438,
+            line_number: 470,
         },
         GrammarRule {
             name: r#"dot_call"#.to_string(),
@@ -765,7 +771,7 @@ pub fn parser_grammar() -> ParserGrammar {
                     ] }) },
                 GrammarElement::Optional { element: Box::new(GrammarElement::RuleReference { name: r#"block"#.to_string() }) },
             ] },
-            line_number: 439,
+            line_number: 471,
         },
         GrammarRule {
             name: r#"scope_resolution"#.to_string(),
@@ -776,19 +782,26 @@ pub fn parser_grammar() -> ParserGrammar {
                         GrammarElement::TokenReference { name: r#"KEYWORD"#.to_string() },
                     ] }) },
             ] },
-            line_number: 447,
+            line_number: 479,
         },
         GrammarRule {
             name: r#"call_arg"#.to_string(),
-            body: GrammarElement::Sequence { elements: vec![
-                GrammarElement::Optional { element: Box::new(GrammarElement::Alternation { choices: vec![
-                        GrammarElement::Literal { value: r#"*"#.to_string() },
-                        GrammarElement::Literal { value: r#"**"#.to_string() },
-                        GrammarElement::Literal { value: r#"&"#.to_string() },
-                    ] }) },
-                GrammarElement::RuleReference { name: r#"expression"#.to_string() },
+            body: GrammarElement::Alternation { choices: vec![
+                GrammarElement::Sequence { elements: vec![
+                    GrammarElement::TokenReference { name: r#"NAME"#.to_string() },
+                    GrammarElement::TokenReference { name: r#"COLON"#.to_string() },
+                    GrammarElement::RuleReference { name: r#"expression"#.to_string() },
+                ] },
+                GrammarElement::Sequence { elements: vec![
+                    GrammarElement::Optional { element: Box::new(GrammarElement::Alternation { choices: vec![
+                            GrammarElement::Literal { value: r#"*"#.to_string() },
+                            GrammarElement::Literal { value: r#"**"#.to_string() },
+                            GrammarElement::Literal { value: r#"&"#.to_string() },
+                        ] }) },
+                    GrammarElement::RuleReference { name: r#"expression"#.to_string() },
+                ] },
             ] },
-            line_number: 477,
+            line_number: 534,
         },
         GrammarRule {
             name: r#"method_call_no_paren"#.to_string(),
@@ -803,17 +816,17 @@ pub fn parser_grammar() -> ParserGrammar {
                         GrammarElement::RuleReference { name: r#"expression"#.to_string() },
                     ] }) },
             ] },
-            line_number: 491,
+            line_number: 548,
         },
         GrammarRule {
             name: r#"expression_stmt"#.to_string(),
             body: GrammarElement::RuleReference { name: r#"expression"#.to_string() },
-            line_number: 492,
+            line_number: 549,
         },
         GrammarRule {
             name: r#"expression"#.to_string(),
             body: GrammarElement::RuleReference { name: r#"ternary"#.to_string() },
-            line_number: 599,
+            line_number: 656,
         },
         GrammarRule {
             name: r#"ternary"#.to_string(),
@@ -826,7 +839,7 @@ pub fn parser_grammar() -> ParserGrammar {
                         GrammarElement::RuleReference { name: r#"expression"#.to_string() },
                     ] }) },
             ] },
-            line_number: 600,
+            line_number: 657,
         },
         GrammarRule {
             name: r#"range"#.to_string(),
@@ -849,7 +862,7 @@ pub fn parser_grammar() -> ParserGrammar {
                         ] }) },
                 ] },
             ] },
-            line_number: 601,
+            line_number: 658,
         },
         GrammarRule {
             name: r#"logical_or"#.to_string(),
@@ -863,7 +876,7 @@ pub fn parser_grammar() -> ParserGrammar {
                         GrammarElement::RuleReference { name: r#"logical_and"#.to_string() },
                     ] }) },
             ] },
-            line_number: 602,
+            line_number: 659,
         },
         GrammarRule {
             name: r#"logical_and"#.to_string(),
@@ -877,7 +890,7 @@ pub fn parser_grammar() -> ParserGrammar {
                         GrammarElement::RuleReference { name: r#"logical_not"#.to_string() },
                     ] }) },
             ] },
-            line_number: 603,
+            line_number: 660,
         },
         GrammarRule {
             name: r#"logical_not"#.to_string(),
@@ -888,7 +901,7 @@ pub fn parser_grammar() -> ParserGrammar {
                         ] }) }) },
                 GrammarElement::RuleReference { name: r#"comparison"#.to_string() },
             ] },
-            line_number: 610,
+            line_number: 667,
         },
         GrammarRule {
             name: r#"comparison"#.to_string(),
@@ -906,7 +919,7 @@ pub fn parser_grammar() -> ParserGrammar {
                         GrammarElement::RuleReference { name: r#"sum"#.to_string() },
                     ] }) },
             ] },
-            line_number: 611,
+            line_number: 668,
         },
         GrammarRule {
             name: r#"sum"#.to_string(),
@@ -920,7 +933,7 @@ pub fn parser_grammar() -> ParserGrammar {
                         GrammarElement::RuleReference { name: r#"term"#.to_string() },
                     ] }) },
             ] },
-            line_number: 612,
+            line_number: 669,
         },
         GrammarRule {
             name: r#"term"#.to_string(),
@@ -934,7 +947,7 @@ pub fn parser_grammar() -> ParserGrammar {
                         GrammarElement::RuleReference { name: r#"factor"#.to_string() },
                     ] }) },
             ] },
-            line_number: 613,
+            line_number: 670,
         },
         GrammarRule {
             name: r#"factor"#.to_string(),
@@ -973,7 +986,7 @@ pub fn parser_grammar() -> ParserGrammar {
                         GrammarElement::RuleReference { name: r#"scope_resolution"#.to_string() },
                     ] }) },
             ] },
-            line_number: 649,
+            line_number: 706,
         },
         GrammarRule {
             name: r#"lambda_literal"#.to_string(),
@@ -986,7 +999,7 @@ pub fn parser_grammar() -> ParserGrammar {
                     ] }) },
                 GrammarElement::RuleReference { name: r#"block"#.to_string() },
             ] },
-            line_number: 668,
+            line_number: 725,
         },
         GrammarRule {
             name: r#"unary_minus"#.to_string(),
@@ -994,7 +1007,7 @@ pub fn parser_grammar() -> ParserGrammar {
                 GrammarElement::TokenReference { name: r#"MINUS"#.to_string() },
                 GrammarElement::RuleReference { name: r#"factor"#.to_string() },
             ] },
-            line_number: 669,
+            line_number: 726,
         },
         GrammarRule {
             name: r#"defined_expression"#.to_string(),
@@ -1002,7 +1015,7 @@ pub fn parser_grammar() -> ParserGrammar {
                 GrammarElement::Literal { value: r#"defined?"#.to_string() },
                 GrammarElement::RuleReference { name: r#"factor"#.to_string() },
             ] },
-            line_number: 680,
+            line_number: 737,
         },
         GrammarRule {
             name: r#"symbol_literal"#.to_string(),
@@ -1014,7 +1027,7 @@ pub fn parser_grammar() -> ParserGrammar {
                         GrammarElement::TokenReference { name: r#"STRING"#.to_string() },
                     ] }) },
             ] },
-            line_number: 687,
+            line_number: 744,
         },
         GrammarRule {
             name: r#"array_literal"#.to_string(),
@@ -1029,7 +1042,7 @@ pub fn parser_grammar() -> ParserGrammar {
                     ] }) },
                 GrammarElement::TokenReference { name: r#"RBRACKET"#.to_string() },
             ] },
-            line_number: 688,
+            line_number: 745,
         },
         GrammarRule {
             name: r#"hash_literal"#.to_string(),
@@ -1044,7 +1057,7 @@ pub fn parser_grammar() -> ParserGrammar {
                     ] }) },
                 GrammarElement::TokenReference { name: r#"RBRACE"#.to_string() },
             ] },
-            line_number: 689,
+            line_number: 746,
         },
         GrammarRule {
             name: r#"hash_entry"#.to_string(),
@@ -1064,7 +1077,7 @@ pub fn parser_grammar() -> ParserGrammar {
                     GrammarElement::RuleReference { name: r#"expression"#.to_string() },
                 ] },
             ] },
-            line_number: 690,
+            line_number: 747,
         },
     ],
         version: 1,